- update libsbml to 5.16.1

+++ src/libSBML/src/sbml/xml/XMLErrorLog.cpp
@@ -7,7 +7,7 @@
  * This file is part of libSBML.  Please visit http://sbml.org for more
  * information about SBML, and the latest version of libSBML.
  *
- * Copyright (C) 2013-2017 jointly by the following organizations:
+ * Copyright (C) 2013-2018 jointly by the following organizations:
  *     1. California Institute of Technology, Pasadena, CA, USA
  *     2. EMBL European Bioinformatics Institute (EMBL-EBI), Hinxton, UK
  *     3. University of Heidelberg, Heidelberg, Germany
@@ -59,8 +59,8 @@ XMLErrorLog::XMLErrorLog ()
 
 /** @cond doxygenLibsbmlInternal */
 /*
-* Copy Constructor
-*/
+ * Copy Constructor
+ */
 XMLErrorLog::XMLErrorLog (const XMLErrorLog& other)
   : mParser(NULL)
   , mOverriddenSeverity(other.mOverriddenSeverity)
@@ -71,8 +71,8 @@ XMLErrorLog::XMLErrorLog (const XMLErrorLog& other)
 
 /** @cond doxygenLibsbmlInternal */
 /*
-* Assignment operator
-*/
+ * Assignment operator
+ */
 XMLErrorLog& XMLErrorLog::operator=(const XMLErrorLog& other)  
 {
   if (this != &other)
@@ -335,7 +335,7 @@ XMLErrorLog::changeErrorSeverity(XMLErrorSeverity_t originalSeverity,
 
   for (iter = mErrors.begin(); iter != mErrors.end(); ++iter) 
   {
-    if ((*iter)->getSeverity() == originalSeverity)
+    if ((*iter)->getSeverity() == (unsigned int)(originalSeverity))
     {
       if (package == "all" || (*iter)->getPackage() == package)
       {
@@ -347,8 +347,8 @@ XMLErrorLog::changeErrorSeverity(XMLErrorSeverity_t originalSeverity,
 }
 
 /*
-* Helper class used by XMLErrorLog::contains.
-*/
+ * Helper class used by XMLErrorLog::contains.
+ */
 class MatchErrorId
 {
 public:
